+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/activities/PhotoVideoActivity.kt
@@ -7,10 +7,11 @@ import android.graphics.drawable.ColorDrawable
 import android.net.Uri
 import android.os.Bundle
 import android.provider.MediaStore
+import android.text.Html
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
-import android.view.WindowManager
+import android.view.WindowInsetsController
 import com.simplemobiletools.commons.dialogs.PropertiesDialog
 import com.simplemobiletools.commons.extensions.*
 import com.simplemobiletools.commons.helpers.*
@@ -37,9 +38,12 @@ open class PhotoVideoActivity : SimpleActivity(), ViewPagerFragment.FragmentList
     var mIsVideo = false
 
     public override fun onCreate(savedInstanceState: Bundle?) {
+        if (config.isUsingSystemTheme) {
+            setTheme(R.style.AppTheme_Material)
+        }
+
         super.onCreate(savedInstanceState)
         setContentView(R.layout.fragment_holder)
-
         if (checkAppSideloading()) {
             return
         }
@@ -51,6 +55,10 @@ open class PhotoVideoActivity : SimpleActivity(), ViewPagerFragment.FragmentList
                 toast(R.string.no_storage_permissions)
                 finish()
             }
+        }
+
+        if (isRPlus()) {
+            window.insetsController?.setSystemBarsAppearance(0, WindowInsetsController.APPEARANCE_LIGHT_STATUS_BARS)
         }
     }
 
@@ -82,7 +90,7 @@ open class PhotoVideoActivity : SimpleActivity(), ViewPagerFragment.FragmentList
             findItem(R.id.menu_show_on_map).isVisible = visibleBottomActions and BOTTOM_ACTION_SHOW_ON_MAP == 0
         }
 
-        updateMenuItemColors(menu)
+        updateMenuItemColors(menu, forceWhiteIcons = true)
         return true
     }
 
@@ -134,7 +142,7 @@ open class PhotoVideoActivity : SimpleActivity(), ViewPagerFragment.FragmentList
         if (intent.extras?.containsKey(REAL_FILE_PATH) == true) {
             val realPath = intent.extras!!.getString(REAL_FILE_PATH)
             if (realPath != null && getDoesFilePathExist(realPath)) {
-                val avoidShowingHiddenFiles = isRPlus() && File(realPath).isHidden
+                val avoidShowingHiddenFiles = isRPlus() && (File(realPath).isHidden || File(realPath.getParentPath(), NOMEDIA).exists())
                 if (!avoidShowingHiddenFiles) {
                     if (realPath.getFilenameFromPath().contains('.') || filename.contains('.')) {
                         if (isFileTypeVisible(realPath)) {
@@ -174,6 +182,7 @@ open class PhotoVideoActivity : SimpleActivity(), ViewPagerFragment.FragmentList
             }
         }
 
+        top_shadow.layoutParams.height = statusBarHeight + actionBarHeight
         checkNotchSupport()
         showSystemUI(true)
         val bundle = Bundle()
@@ -190,7 +199,7 @@ open class PhotoVideoActivity : SimpleActivity(), ViewPagerFragment.FragmentList
 
         mIsVideo = type == TYPE_VIDEOS
         mMedium = Medium(null, filename, mUri.toString(), mUri!!.path!!.getParentPath(), 0, 0, file.length(), type, 0, false, 0L, 0)
-        supportActionBar?.title = mMedium!!.name
+        supportActionBar?.title = Html.fromHtml("<font color='${Color.WHITE.toHex()}'>${mMedium!!.name}</font>")
         bundle.putSerializable(MEDIUM, mMedium)
 
         if (savedInstanceState == null) {