+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/activities/WidgetConfigureActivity.kt
@@ -48,15 +48,19 @@ class WidgetConfigureActivity : SimpleActivity() {
         config_text_color.setOnClickListener { pickTextColor() }
         folder_picker_value.setOnClickListener { changeSelectedFolder() }
         config_image_holder.setOnClickListener { changeSelectedFolder() }
+
+        updateTextColors(folder_picker_holder)
+        val primaryColor = getProperPrimaryColor()
+        config_bg_seekbar.setColors(mTextColor, primaryColor, primaryColor)
+        config_save.setTextColor(mTextColor)
+        folder_picker_holder.background = ColorDrawable(getProperBackgroundColor())
+
         folder_picker_show_folder_name.isChecked = config.showWidgetFolderName
         handleFolderNameDisplay()
         folder_picker_show_folder_name_holder.setOnClickListener {
             folder_picker_show_folder_name.toggle()
             handleFolderNameDisplay()
         }
-
-        updateTextColors(folder_picker_holder)
-        folder_picker_holder.background = ColorDrawable(config.backgroundColor)
 
         getCachedDirectories(false, false) {
             mDirectories = it
@@ -122,15 +126,13 @@ class WidgetConfigureActivity : SimpleActivity() {
 
     private fun updateBackgroundColor() {
         mBgColor = mBgColorWithoutTransparency.adjustAlpha(mBgAlpha)
-        config_save.setBackgroundColor(mBgColor)
         config_image_holder.setBackgroundColor(mBgColor)
-        config_bg_color.setFillWithStroke(mBgColor, Color.BLACK)
+        config_bg_color.setFillWithStroke(mBgColor, mBgColor)
     }
 
     private fun updateTextColor() {
-        config_save.setTextColor(mTextColor)
         config_folder_name.setTextColor(mTextColor)
-        config_text_color.setFillWithStroke(mTextColor, Color.BLACK)
+        config_text_color.setFillWithStroke(mTextColor, mTextColor)
     }
 
     private fun pickBackgroundColor() {
@@ -152,7 +154,7 @@ class WidgetConfigureActivity : SimpleActivity() {
     }
 
     private fun changeSelectedFolder() {
-        PickDirectoryDialog(this, "", false, true) {
+        PickDirectoryDialog(this, "", false, true, false, true) {
             updateFolderImage(it)
         }
     }