+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/extensions/Activity.kt
@@ -15,8 +15,10 @@ import android.os.Build
 import android.provider.MediaStore
 import android.provider.MediaStore.Files
 import android.provider.MediaStore.Images
+import android.provider.Settings
 import android.util.DisplayMetrics
 import android.view.View
+import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
 import androidx.exifinterface.media.ExifInterface
 import com.bumptech.glide.Glide
@@ -121,6 +123,17 @@ fun SimpleActivity.launchAbout() {
     }
 
     startAboutActivity(R.string.app_name, licenses, BuildConfig.VERSION_NAME, faqItems, true)
+}
+
+
+fun BaseSimpleActivity.handleMediaManagementPrompt(callback: () -> Unit) {
+    if (isSPlus() && !MediaStore.canManageMedia(this)) {
+        ConfirmationDialog(this, "", R.string.media_management_prompt, R.string.ok, 0) {
+            launchMediaManagementIntent(callback)
+        }
+    } else {
+        callback()
+    }
 }
 
 fun AppCompatActivity.showSystemUI(toggleActionBarVisibility: Boolean) {
@@ -246,7 +259,7 @@ fun BaseSimpleActivity.tryCopyMoveFilesTo(fileDirItems: ArrayList<FileDirItem>, 
     }
 
     val source = fileDirItems[0].getParentPath()
-    PickDirectoryDialog(this, source, true, false) {
+    PickDirectoryDialog(this, source, true, false, true, false) {
         val destination = it
         handleSAFDialog(source) {
             if (it) {
@@ -341,13 +354,36 @@ fun BaseSimpleActivity.restoreRecycleBinPath(path: String, callback: () -> Unit)
 fun BaseSimpleActivity.restoreRecycleBinPaths(paths: ArrayList<String>, callback: () -> Unit) {
     ensureBackgroundThread {
         val newPaths = ArrayList<String>()
+        var shownRestoringToPictures = false
         for (source in paths) {
-            val destination = source.removePrefix(recycleBinPath)
+            var destination = source.removePrefix(recycleBinPath)
+
+            val destinationParent = destination.getParentPath()
+            if (isRestrictedWithSAFSdk30(destinationParent) && !isInDownloadDir(destinationParent)) {
+                // if the file is not writeable on SDK30+, change it to Pictures
+                val picturesDirectory = getPicturesDirectoryPath(destination)
+                destination = File(picturesDirectory, destination.getFilenameFromPath()).path
+                if (!shownRestoringToPictures) {
+                    toast(getString(R.string.restore_to_path, humanizePath(picturesDirectory)))
+                    shownRestoringToPictures = true
+                }
+            }
+
             val lastModified = File(source).lastModified()
 
             val isShowingSAF = handleSAFDialog(destination) {}
             if (isShowingSAF) {
                 return@ensureBackgroundThread
+            }
+
+            val isShowingSAFSdk30 = handleSAFDialogSdk30(destination) {}
+            if (isShowingSAFSdk30) {
+                return@ensureBackgroundThread
+            }
+
+            if (getDoesFilePathExist(destination)) {
+                val newFile = getAlternativeFile(File(destination))
+                destination = newFile.path
             }
 
             var inputStream: InputStream? = null
@@ -368,7 +404,7 @@ fun BaseSimpleActivity.restoreRecycleBinPaths(paths: ArrayList<String>, callback
                 out?.flush()
 
                 if (File(source).length() == copiedSize) {
-                    mediaDB.updateDeleted(destination.removePrefix(recycleBinPath), 0, "$RECYCLE_BIN$destination")
+                    mediaDB.updateDeleted(destination.removePrefix(recycleBinPath), 0, "$RECYCLE_BIN${source.removePrefix(recycleBinPath)}")
                 }
                 newPaths.add(destination)
 
@@ -450,7 +486,7 @@ fun AppCompatActivity.fixDateTaken(
     callback: (() -> Unit)? = null
 ) {
     val BATCH_SIZE = 50
-    if (showToasts) {
+    if (showToasts && !hasRescanned) {
         toast(R.string.fixing)
     }