+++ CHANGELOG.md
@@ -1,6 +1,29 @@
 Changelog
 ==========
 
+Version 6.23.2 *(2022-05-01)*
+----------------------------
+
+ * Simplify some file operations on Android 12+, if the gallery is set as a Media Management app
+ * Added .avif support on Android 12
+ * Allow setting both landscape and portrait wallpapers
+ * Added some translation, stability and UX improvements
+
+Version 6.23.1 *(2022-04-16)*
+----------------------------
+
+ * Fixed some file operations on Android 10 and 11
+ * Restore Recycle bin files in Pictures, if the original folder is inaccessible
+
+Version 6.23.0 *(2022-04-15)*
+----------------------------
+
+ * Mark favorited file thumbnails
+ * Added an Auto and Material You theme on Android 12+
+ * Changed some file operations to comply with the latest Google permissions policy, apps like this can no longer show hidden items on Android 11+
+ * Properly delete items in the recycle bin after 30 days
+ * Added many other stability, translation, performance, UX and UI improvements
+
 Version 6.22.3 *(2021-12-19)*
 ----------------------------